Refactor field_path into schema package
