+++ src/gem/schema/field_path/ops.py
@@ -1,149 +1,19 @@
-"""Huffman-coded field path decoder for Dota 2 Source 2 entity delta streams.
+"""Manta-compatible field-path operation table.
 
-Field paths describe which fields of an entity have changed.  Each path is a
-list of up to 7 integer indices (one per nesting level) that navigate from
-a serializer root to a leaf field.  The sequence of paths in a packet is
-Huffman-encoded using the operation weights defined below.
-
-Reference: manta/field_path.go, manta/huffman.go
+The operation order and weights must stay byte-for-byte compatible with
+``refs/manta/field_path.go`` because Huffman symbol values are table indices.
 """
 
 from __future__ import annotations
 
-import heapq
-import struct
 from collections.abc import Callable
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
+
+from gem.schema.field_path.model import FieldPath
 
 if TYPE_CHECKING:
     from gem.binary.reader import BitReader
-
-# ---------------------------------------------------------------------------
-# FieldPath
-# ---------------------------------------------------------------------------
-
-_RESET = [-1, 0, 0, 0, 0, 0, 0]
-
-
-class FieldPath:
-    """A mutable path of up to 7 integer field indices.
-
-    Starts at ``path = [-1, 0, 0, 0, 0, 0, 0]``, ``last = 0``.
-    Operations mutate ``path`` and ``last`` in place.
-
-    Attributes:
-        path: Seven-element list; active indices are ``path[0:last+1]``.
-        last: Index of the deepest active level (0-based).
-        done: Set to True by ``FieldPathEncodeFinish`` to stop iteration.
-    """
-
-    __slots__ = ("path", "last", "done")
-
-    def __init__(self) -> None:
-        self.path: list[int] = [-1, 0, 0, 0, 0, 0, 0]
-        self.last: int = 0
-        self.done: bool = False
-
-    def reset(self) -> None:
-        """Reset to the initial empty state."""
-        self.path[:] = _RESET
-        self.last = 0
-        self.done = False
-
-    def pop(self, n: int) -> None:
-        """Pop n levels off the path, zeroing the vacated slots.
-
-        Args:
-            n: Number of levels to remove.
-        """
-        for _ in range(n):
-            self.path[self.last] = 0
-            self.last -= 1
-
-    def copy(self) -> FieldPath:
-        """Return an independent copy of this path.
-
-        Returns:
-            A new FieldPath with identical state.
-        """
-        fp = FieldPath()
-        fp.path[:] = self.path
-        fp.last = self.last
-        fp.done = self.done
-        return fp
-
-    def to_tuple(self) -> tuple[int, ...]:
-        """Return the active indices as an immutable tuple.
-
-        Returns:
-            Tuple of active integer indices, e.g. ``(2, 0, 5)``.
-        """
-        return tuple(self.path[: self.last + 1])
-
-    def to_str(self) -> str:
-        """Return a slash-separated string of active indices.
-
-        Returns:
-            String like ``"2/0/5"``.
-        """
-        return "/".join(str(self.path[i]) for i in range(self.last + 1))
-
-    def plus_one(self) -> None:
-        """Increment the deepest index by 1."""
-        self.path[self.last] += 1
-
-
-# ---------------------------------------------------------------------------
-# Huffman tree
-# ---------------------------------------------------------------------------
-
-
-@dataclass(order=True)
-class _HNode:
-    """Internal min-heap node for Huffman tree construction."""
-
-    weight: int
-    value: int
-    left: _HNode | None = field(default=None, compare=False)
-    right: _HNode | None = field(default=None, compare=False)
-
-    @property
-    def is_leaf(self) -> bool:
-        return self.left is None and self.right is None
-
-
-def _build_huffman_tree(weights: list[int]) -> _HNode:
-    """Build a Huffman tree from a list of symbol weights.
-
-    Symbols with weight 0 are treated as weight 1 (must appear in tree).
-    Tie-breaking: higher value wins (matches manta's Go implementation).
-
-    Args:
-        weights: Per-symbol frequency weights, indexed by symbol value.
-
-    Returns:
-        Root _HNode of the completed Huffman tree.
-    """
-    heap: list[tuple[int, int, _HNode]] = []
-    for v, w in enumerate(weights):
-        w = w or 1
-        heapq.heappush(heap, (w, -v, _HNode(w, v)))
-
-    counter = len(weights)
-    while len(heap) > 1:
-        wa, _, a = heapq.heappop(heap)
-        wb, _, b = heapq.heappop(heap)
-        merged = _HNode(wa + wb, counter, left=a, right=b)
-        heapq.heappush(heap, (merged.weight, -counter, merged))
-        counter += 1
-
-    return heap[0][2]
-
-
-# ---------------------------------------------------------------------------
-# Field path operation table (40 ops, weights match manta exactly)
-# ---------------------------------------------------------------------------
 
 
 @dataclass(frozen=True)
@@ -422,133 +292,3 @@ def _make_ops() -> list[FieldPathOp]:
 
 
 FIELD_PATH_OPS: list[FieldPathOp] = _make_ops()
-HUFF_TREE: _HNode = _build_huffman_tree([op.weight for op in FIELD_PATH_OPS])
-
-
-# ---------------------------------------------------------------------------
-# Flat Huffman decode table
-# ---------------------------------------------------------------------------
-
-
-def _build_decode_table(root: _HNode, table_bits: int) -> list[tuple[int, int]]:
-    """Build a flat O(1) decode table from the Huffman tree.
-
-    Each entry ``table[i] = (op_index, bits_consumed)`` covers all ``table_bits``-
-    bit integers whose leading bits match the Huffman code for ``op_index``.
-    Shorter codes fill multiple entries (one per possible suffix).
-
-    Args:
-        root: Root of the Huffman tree.
-        table_bits: Width of the table in bits (``2**table_bits`` entries).
-
-    Returns:
-        List of ``(op_index, bits_consumed)`` tuples, indexed by the
-        ``table_bits``-bit integer peeked from the bit stream.
-    """
-    size = 1 << table_bits
-    table: list[tuple[int, int]] = [(0, 0)] * size
-
-    stack: list[tuple[_HNode, int, int]] = [(root, 0, 0)]  # node, code, depth
-    while stack:
-        node, code, depth = stack.pop()
-        if node.is_leaf:
-            # Fill all entries whose top `depth` bits match `code`
-            suffix_count = 1 << (table_bits - depth)
-            for s in range(suffix_count):
-                table[code | (s << depth)] = (node.value, depth)
-        else:
-            if node.left is not None:
-                stack.append((node.left, code, depth + 1))
-            if node.right is not None:
-                stack.append((node.right, code | (1 << depth), depth + 1))
-
-    return table
-
-
-def _tree_depth(node: _HNode) -> int:
-    """Return the maximum depth of the Huffman tree.
-
-    Args:
-        node: Root node.
-
-    Returns:
-        int: Maximum leaf depth (root = depth 0).
-    """
-    if node.is_leaf:
-        return 0
-    left_d = _tree_depth(node.left) + 1 if node.left else 0
-    right_d = _tree_depth(node.right) + 1 if node.right else 0
-    return max(left_d, right_d)
-
-
-_HUFF_TABLE_BITS: int = _tree_depth(HUFF_TREE)
-_HUFF_DECODE_TABLE: list[tuple[int, int]] = _build_decode_table(HUFF_TREE, _HUFF_TABLE_BITS)
-
-
-# ---------------------------------------------------------------------------
-# Field path reading
-# ---------------------------------------------------------------------------
-
-
-def read_field_paths(r: BitReader) -> list[FieldPath]:
-    """Decode a Huffman-coded sequence of field paths from r.
-
-    Uses a flat O(1) decode table to resolve each Huffman op in a single
-    peek + skip rather than a per-bit tree walk.  The peek/skip/rem_bits
-    calls are inlined directly against BitReader's private attributes to
-    eliminate ~22M Python function calls per replay.
-
-    Falls back to the tree walk for the last few bits when fewer than
-    ``_HUFF_TABLE_BITS`` bits remain in the buffer.
-
-    Args:
-        r: BitReader positioned at the start of the field path sequence.
-
-    Returns:
-        List of FieldPath objects, one per updated field (not including the
-        finish sentinel).
-    """
-    fp = FieldPath()
-    paths: list[FieldPath] = []
-    ops = FIELD_PATH_OPS
-    table = _HUFF_DECODE_TABLE
-    table_bits = _HUFF_TABLE_BITS
-    mask = (1 << table_bits) - 1
-
-    buf = r._buf
-    size = r._size
-    unpack_from = struct.unpack_from
-
-    while not fp.done:
-        # Inline rem_bits: (size - pos) * 8 + bit_count
-        if (size - r._pos) * 8 + r._bit_count >= table_bits:
-            # Inline peek_bits(table_bits): refill then read without consuming
-            while table_bits > r._bit_count:
-                remaining = size - r._pos
-                if remaining >= 4:
-                    r._bit_val |= unpack_from("<I", buf, r._pos)[0] << r._bit_count
-                    r._pos += 4
-                    r._bit_count += 32
-                elif remaining > 0:
-                    r._bit_val |= buf[r._pos] << r._bit_count
-                    r._pos += 1
-                    r._bit_count += 8
-                else:
-                    break
-            bits = r._bit_val & mask
-            op_idx, consumed = table[bits]
-            # Inline skip_bits(consumed)
-            r._bit_val >>= consumed
-            r._bit_count -= consumed
-        else:
-            # Fallback: tree walk for the last few bits
-            node = HUFF_TREE
-            while not node.is_leaf:
-                node = node.right if r.read_bits(1) else node.left  # type: ignore[assignment]
-            op_idx = node.value
-
-        ops[op_idx].fn(r, fp)
-        if not fp.done:
-            paths.append(fp.copy())
-
-    return paths
